# api/openapi.yaml
openapi: 3.0.3
info: { title: Cerply Core API, version: "0.1.0" }
paths:
...
  /api/certified/verify:
    post:
      summary: Verify artifact signature and integrity (extended)
      description: |
        Supports three verification modes:
        1. By artifactId: { "artifactId": "..." }
        2. Inline artifact + signature: { "artifact": {...}, "signature": "base64" }
        3. Legacy plan lock: { "plan": {...}, "lock": {...} }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              oneOf:
                - required: [artifactId]
                  properties:
                    artifactId: { type: string }
                    meta: { type: object, properties: { request_id: { type: string } } }
                - required: [artifact, signature]
                  properties:
                    artifact:
                      type: object
                      description: "Artifact JSON to verify"
                    signature:
                      type: string
                      description: "Base64-encoded Ed25519 signature"
                    meta: { type: object, properties: { request_id: { type: string } } }
                - required: [plan, lock]
                  properties:
                    plan: { type: object }
                    lock:
                      type: object
                      required: [algo, hash]
                      properties:
                        algo: { type: string, enum: [blake3, sha256] }
                        hash: { type: string }
                        canonical_bytes: { type: integer }
                    meta: { type: object, properties: { request_id: { type: string } } }
      responses:
        '200':
          description: Verification result
          headers:
            Access-Control-Allow-Origin:
              schema: { type: string, enum: ['*'] }
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [ok]
                    properties:
                      ok: { type: boolean, enum: [true] }
                      artifactId: { type: string }
                      sha256: { type: string }
                      lockHash: { type: string }
                  - type: object
                    required: [ok, reason]
                    properties:
                      ok: { type: boolean, enum: [false] }
                      reason:
                        type: string
                        enum: [lock_mismatch, sha256_mismatch, signature_invalid]
                      details: { type: object }
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: object
                    required: [code, message]
                    properties:
                      code: { type: string, enum: [BAD_REQUEST] }
                      message: { type: string }
        '404':
          description: Artifact not found
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: object
                    required: [code, message]
                    properties:
                      code: { type: string, enum: [NOT_FOUND, FILE_NOT_FOUND] }
                      message: { type: string }
        '413':
          description: Payload too large (exceeds 16KB)
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error:
                    type: object
                    required: [code, message, details]
                    properties:
                      code: { type: string, enum: [PAYLOAD_TOO_LARGE] }
                      message: { type: string }
                      details:
                        type: object
                        required: [limit, size]
                        properties:
                          limit: { type: integer }
                          size: { type: integer }
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token